requirements.txt
# Core ML & Audio Processing
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
librosa>=0.10.0
soundfile>=0.12.0
pyannote.audio>=3.0.0
# Note: silero-vad is loaded via torch.hub, no pip package needed

# Video Processing
opencv-python>=4.8.0
mediapipe>=0.10.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
pydantic>=2.0.0

# Development (optional)
pytest>=7.4.0
jupyter>=1.0.0
black>=23.0.0
